requirements.txt
asgiref==3.7.2
Django==3.2
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.28
phonenumberslite==8.13.28
pillow==10.2.0
PyJWT==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
